Update uninstall.php to conditionally delete options
Modify uninstall process to conditionally delete options based on user preference.

+++ uninstall.php
@@ -9,6 +9,13 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-delete_option( 'ayudawp_blg_settings' );
-delete_option( 'ayudawp_blg_dns_cache' );
-delete_option( 'ayudawp_blg_bypass_state' );
+$config = get_option( 'ayudawp_blg_settings', array() );
+
+/* Only delete data if the user opted in (default: yes) */
+$delete = isset( $config['delete_data'] ) ? (int) $config['delete_data'] : 1;
+
+if ( $delete ) {
+	delete_option( 'ayudawp_blg_settings' );
+	delete_option( 'ayudawp_blg_dns_cache' );
+	delete_option( 'ayudawp_blg_bypass_state' );
+}
